add tests for reviewer map and inbox services; remove unused tests
